+++ legacy/tests/integration/fundTrading.js
@@ -1,13 +1,19 @@
 import test from "ava";
 import web3 from "../../utils/lib/web3";
-import api from "../../utils/lib/api";
 import { deployContract, retrieveContract } from "../../utils/lib/contracts";
 import getAllBalances from "../../utils/lib/getAllBalances";
 import deployEnvironment from "../../utils/deploy/contracts";
 import { getTermsSignatureParameters } from "../../utils/lib/signing";
 import getFundComponents from "../../utils/lib/getFundComponents";
 import { updateTestingPriceFeed } from "../../utils/lib/updatePriceFeed";
-import { makeOrderSignature, takeOrderSignature, cancelOrderSignature } from "../../utils/lib/data";
+import {
+  makeOrderSignature,
+  takeOrderSignature,
+  cancelOrderSignature,
+  makeOrderSignatureString,
+  takeOrderSignatureString,
+  cancelOrderSignatureString,
+} from "../../utils/lib/data";
 
 const BigNumber = require("bignumber.js");
 const environmentConfig = require("../../utils/config/environment.js");
@@ -120,12 +126,12 @@ test.beforeEach(async () => {
 const initialTokenAmount = new BigNumber(10 ** 23);
 test.serial("investor receives initial ethToken for testing", async t => {
   const pre = await getAllBalances(deployed, accounts, fund);
-  const preDeployerEth = new BigNumber(await api.eth.getBalance(deployer));
+  const preDeployerEth = new BigNumber(await web3.eth.getBalance(deployer));
   receipt = await ethToken.methods.transfer(investor, initialTokenAmount.toFixed()).send(
     { from: deployer, gasPrice: config.gasPrice }
   );
   runningGasTotal = runningGasTotal.plus(receipt.gasUsed);
-  const postDeployerEth = new BigNumber(await api.eth.getBalance(deployer));
+  const postDeployerEth = new BigNumber(await web3.eth.getBalance(deployer));
   const post = await getAllBalances(deployed, accounts, fund);
 
   t.deepEqual(
@@ -231,11 +237,11 @@ exchangeIndexes.forEach(i => {
     await updateTestingPriceFeed(deployed);
     receipt = await fund.trading.methods.callOnExchange(
       i,
-      makeOrderSignature,
-      [web3.utils.randomHex(20), web3.utils.randomHex(20), ethToken.options.address, mlnToken.options.address, web3.utils.randomHex(20)],
+      makeOrderSignatureString,
+      [web3.utils.randomHex(20), web3.utils.randomHex(20), ethToken.options.address, mlnToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
       [trade1.sellQuantity.toFixed(), trade1.buyQuantity.toFixed(), 0, 0, 0, 0, 0, 0],
       web3.utils.padLeft('0x0', 64),
-      0,
+      web3.utils.padLeft('0x0', 64),
       web3.utils.padLeft('0x0', 64),
       web3.utils.padLeft('0x0', 64),
     ).send({ from: manager, gas: config.gas, gasPrice: config.gasPrice });
@@ -385,11 +391,11 @@ exchangeIndexes.forEach(i => {
       const orderId = await exchanges[i].methods.last_offer_id().call();
       receipt = await fund.trading.methods.callOnExchange(
         i,
-        takeOrderSignature,
-        [web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
+        takeOrderSignatureString,
+        [web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
         [0, 0, 0, 0, 0, 0, trade2.buyQuantity.toFixed(), 0],
         `0x${Number(orderId).toString(16).padStart(64, "0")}`,
-        0,
+        web3.utils.padLeft('0x0', 64),
         web3.utils.padLeft('0x0', 64),
         web3.utils.padLeft('0x0', 64),
       ).send(
@@ -443,11 +449,11 @@ test.serial(
     const preOrderId = await exchanges[0].methods.last_offer_id().call();
     receipt = await t.throws(fund.trading.methods.callOnExchange(
       0,
-      makeOrderSignature,
-      [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, ethToken.options.address, web3.utils.randomHex(20)],
+      makeOrderSignatureString,
+      [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, ethToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
       [trade3.sellQuantity.toFixed(), trade3.buyQuantity.toFixed(), 0, 0, 0, 0, 0, 0],
       web3.utils.padLeft('0x0', 64),
-      0,
+      web3.utils.padLeft('0x0', 64),
       web3.utils.padLeft('0x0', 64),
       web3.utils.padLeft('0x0', 64),
     ).send(
@@ -527,14 +533,14 @@ test.serial(
 
     await t.throws(fund.trading.methods.callOnExchange(
       0,
-      takeOrderSignature,
-      [web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
-      [0, 0, 0, 0, 0, 0, trade4.buyQuantity.toFixed(), 0],
+      takeOrderSignatureString,
+      [web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
+      [0, 0, 0, 0, 0, 0, trade4.buyQuantity.toFixed(), 0],  // changed from [0, trade4.sellQuantity, 0, 0, 0, 0, 0, 0],
       `0x${Number(orderId)
         .toString(16)
         .padStart(64, "0")}`,
-      0,
-      web3.utils.padLeft('0x0', 64),
+        web3.utils.padLeft('0x0', 64),
+        web3.utils.padLeft('0x0', 64),
       web3.utils.padLeft('0x0', 64),
     ).send(
       { from: manager, gas: config.gas }
@@ -573,11 +579,11 @@ test.serial("manager makes an order and cancels it", async t => {
   runningGasTotal = runningGasTotal.plus(receipt.gasUsed);
   receipt = await fund.trading.methods.callOnExchange(
     0,
-    makeOrderSignature,
-    [web3.utils.randomHex(20), web3.utils.randomHex(20), ethToken.options.address, mlnToken.options.address, web3.utils.randomHex(20)],
+    makeOrderSignatureString,
+    [web3.utils.randomHex(20), web3.utils.randomHex(20), ethToken.options.address, mlnToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
     [trade1.sellQuantity.toFixed(), trade1.buyQuantity.toFixed(), 0, 0, 0, 0, 0, 0],
     web3.utils.padLeft('0x0', 64),
-    0,
+    web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
   ).send({from: manager, gas: config.gas, gasPrice: config.gasPrice});
@@ -586,13 +592,13 @@ test.serial("manager makes an order and cancels it", async t => {
   
   receipt = await fund.trading.methods.callOnExchange(
     0,
-    cancelOrderSignature,
-    [web3.utils.randomHex(20), web3.utils.randomHex(20), ethToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
+    cancelOrderSignatureString,
+    [web3.utils.randomHex(20), web3.utils.randomHex(20), ethToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
     [0, 0, 0, 0, 0, 0, 0, 0],
     `0x${Number(offerNumber)
       .toString(16)
       .padStart(64, "0")}`,
-    0,
+    web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
   ).send({from: manager, gas: config.gas, gasPrice: config.gasPrice});
@@ -782,11 +788,11 @@ test.serial("manger opens new order, but not anyone can cancel", async t => {
   const pre = await getAllBalances(deployed, accounts, fund);
   await fund.trading.methods.callOnExchange(
     0,
-    makeOrderSignature,
-    [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, ethToken.options.address, web3.utils.randomHex(20)],
+    makeOrderSignatureString,
+    [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, ethToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
     [trade1.sellQuantity.toFixed(), trade1.buyQuantity.toFixed(), 0, 0, 0, 0, 0, 0],
     web3.utils.padLeft('0x0', 64),
-    0,
+    web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
   ).send(
@@ -795,13 +801,13 @@ test.serial("manger opens new order, but not anyone can cancel", async t => {
   const offerNumber = await exchanges[0].methods.last_offer_id().call();
   await t.throws(fund.trading.methods.callOnExchange(
     0,
-    cancelOrderSignature,
-    [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
+    cancelOrderSignatureString,
+    [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
     [0, 0, 0, 0, 0, 0, 0, 0],
     `0x${Number(offerNumber)
       .toString(16)
       .padStart(64, "0")}`,
-    0,
+    web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
   ).send(
@@ -852,13 +858,13 @@ test.serial("shutdown of fund allows anyone to cancel order", async t => {
   const offerNumber = await exchanges[0].methods.last_offer_id().call();
   receipt = await fund.trading.methods.callOnExchange(
     0,
-    cancelOrderSignature,
-    [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20)],
+    cancelOrderSignatureString,
+    [web3.utils.randomHex(20), web3.utils.randomHex(20), mlnToken.options.address, web3.utils.randomHex(20), web3.utils.randomHex(20), web3.utils.randomHex(20)],
     [0, 0, 0, 0, 0, 0, 0, 0],
     `0x${Number(offerNumber)
       .toString(16)
       .padStart(64, "0")}`,
-    0,
+    web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
     web3.utils.padLeft('0x0', 64),
   ).send(