+++ legacy/utils/lib/mocks.js
@@ -12,9 +12,7 @@ function newMockBytes4() {
   return web3.utils.randomHex(4);
 }
 
-function zeroAddress() {
-  return web3.utils.toChecksumAddress(`0x${'0'.repeat(20)}`);
-}
+const zeroAddress = web3.utils.toChecksumAddress(`0x${'0'.repeat(40)}`);
 
 export {
   newMockAddress,